Move student deletion to courseService
